Commit 21
-Separated POak class into cpp/hpp files
-Separated GameLoop Class into cpp/hpp files
