+++ Pokemon/Game.hpp
@@ -1,0 +1,16 @@
+#ifndef GAME_HPP
+#define GAME_HPP
+#pragma once
+
+class Player;
+
+class Game
+{
+public:
+    Game();
+    void gameLoop(Player player);   
+};
+
+
+
+#endif
