+++ Pokemon/Pokemon/x64/Debug/Pokemon.tlog/Cl.items.tlog
@@ -1,4 +1,6 @@
+C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Game.cpp;C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\Game.obj
 C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\main.cpp;C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\main.obj
 C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Player.cpp;C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\Player.obj
 C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon.cpp;C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\Pokemon.obj
+C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\ProfessorOak.cpp;C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\ProfessorOak.obj
 C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Utility.cpp;C:\Users\Casey\Documents\GitHub\Pokemon\Pokemon\Pokemon\x64\Debug\Utility.obj
